Replace URI with Addressable::URI
URI's parsing sucks. It raises errors on spaces, UTF-8 characters, and other things that browsers are just fine with.

You'd think that URI.escape would help, but it's not really useful either. It will escape things like '#', and break some URLs. It also forces encoding to US-ASCII for some weird reason.

This is a good enough reason for me to add a dependency on Addressable, which fixes these problems.